Remove possibility that best solutions die
It used to be possible for nondominated solutions to loose a fight and
die as a result. Now, if an agent was never dominated and looses by
crowding comparison, he cannot die as a result. This prevents desolate
good agents from dying due to this seemingly unlikely, but surprisingly
common circumstance.